feat: Add commandto create ELPA project
